requirements.txt
flask
gunicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-secret-manager
openai